week 4 homework
How difficult did you find this (out of 10)? 0 being no problems at all, 10 being impossible; 3
Was there anything that you struggled with?; using the right if statment
Is there anything that you'd like some further information on?; how to deal with data and statments in reagrds to this task
Roughly how long did it take? 3hrs